修复编译警告和错误
【Gradle 配置修复】
- richeditor-compose: 添加 withHostTest {} 启用 Android host tests
- richeditor-compose: 移除冗余的 commonTest dependsOn 配置
- richeditor-compose: 将废弃的 'by getting' 语法替换为 getByName()

【编译警告修复】
- HapticFeedbackController: 注解目标明确指定 @param:ApplicationContext
- SolarTermManager: 移除非空类型上的不必要安全调用
- ImageExporter: 修复类型不匹配（Int → Float）和冗余转换
- ShareCardComponent/TodoListItem/NotificationHelper/BootReceiver: 移除不必要非空断言
- AchievementChecker/ReminderActionReceiver/TimelineInspirationItem: 移除不必要安全调用
- InspirationMenuDropdown/RichTextFormatToolbar/InspirationStatsScreen/StatsScreen: 使用 AutoMirrored 图标
- ZonedReorderableLazyColumn/HomeScreen: 移除冗余 else 分支
- InspirationEditViewModel/TodoEditViewModel: 移除冗余 Elvis 表达式和不必要 cast
- LocationViewModel: 添加 @Deprecated 注解标记覆盖废弃成员
- HomeViewModel/HomeScreen: 移除 contentFormat 上的不必要安全调用

【API 升级】
- InspirationViewScreen: 从 LocalClipboardManager 迁移到 LocalClipboard API
- InspirationScreenshot: 从废弃的 resume() 迁移到 resumeWith(Result.success())

【其他】
- 更新报错汇总.md

+++ app/src/main/java/com/corgimemo/app/ui/components/FloatingCorgiButton.kt
@@ -216,8 +216,8 @@ fun FloatingCorgiButton(
             paddingStartDp = targetStart
 
             // 保存位置（百分比坐标）
-            val percentX = targetStart.value.toFloat() / screenWidthDp.value
-            val percentY = 1f - (paddingBottomDp.value.toFloat() / screenHeightDp.value)
+            val percentX = targetStart.value / screenWidthDp.value
+            val percentY = 1f - (paddingBottomDp.value / screenHeightDp.value)
             onPositionChanged(percentX, percentY)
         }
     }
